fix easing demo and move easings back out of module (#130)
