+++ index.js
@@ -6,7 +6,6 @@
 
 const raylib = require('./src/generated/node-raylib.js')
 const { format } = require('util')
-const easings = require('./src/easings.js')
 
 raylib.MAX_GAMEPADS = 4
 raylib.MAX_GAMEPAD_AXIS = 8
@@ -20,8 +19,6 @@ raylib.MAX_KEY_PRESSED_QUEUE = 16
  * Text formatting with variables (sprintf style)
  */
 raylib.TextFormat = format
-
-raylib.Easings = easings
 
 // Export the bindings for the module.
 module.exports = raylib
